fluffos-2.08

+++ src/call_out.c
@@ -421,7 +421,7 @@ int print_call_out_usage (outbuffer_t * ob, int verbose)
       outbuf_addv(ob, "call out:\t\t\t%8d %8d (current length %d)\n", num_call,
       num_call * sizeof(pending_call_t), i);
     }
-    return (int) (num_call * sizeof(pending_call_t));
+    return num_call * sizeof(pending_call_t);
 }
 
 #ifdef DEBUGMALLOC_EXTENSIONS
