+++ src/compiler.c
@@ -2448,7 +2448,7 @@ static program_t *epilog (void) {
 #ifdef DEBUG
     if (p - (char *)prog != size) {
         fprintf(stderr, "Program size miscalculated for /%s.\n", prog->filename);
-        fprintf(stderr, "is: %i, expected: %i\n", p-(char *)prog, (int) size);
+        fprintf(stderr, "is: %i, expected: %i\n", p-(char *)prog, size);
     }
 #endif
 #ifdef BINARIES
